+++ src/modules/Wallet/pages/Transactions/modals/ModifyTransactionsModal/components/AssetsSelector.tsx
@@ -17,14 +17,6 @@ function AssetsSelector({
 }) {
   const { assets } = useWalletContext()
 
-  if (assets === 'loading') {
-    return <div>Loading...</div>
-  }
-
-  if (assets === 'error') {
-    return <div>Error</div>
-  }
-
   return (
     <ListboxOrComboboxInput
       required